+++ packages/file-explorer/src/FileList.tsx
@@ -1,14 +1,6 @@
-import {
-  ItemList,
-  Range,
-  RenderItemProps,
-  Tooltip,
-} from '@deephaven/components';
-import { dhPython, vsCode, vsFolder, vsFolderOpened } from '@deephaven/icons';
+import { ItemList, Range } from '@deephaven/components';
 import Log from '@deephaven/log';
 import { RangeUtils } from '@deephaven/utils';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 import classNames from 'classnames';
 import React, {
   useCallback,
@@ -19,12 +11,8 @@ import React, {
 } from 'react';
 import { FileStorageItem, FileStorageTable, isDirectory } from './FileStorage';
 import './FileList.scss';
-import FileUtils, { MIME_TYPE } from './FileUtils';
-import {
-  DEFAULT_ROW_HEIGHT,
-  getMoveOperation,
-  getPathFromItem,
-} from './FileListUtils';
+import { DEFAULT_ROW_HEIGHT, getMoveOperation } from './FileListUtils';
+import { FileListItem, FileListRenderItemProps } from './FileListItem';
 
 const log = Log.module('FileList');
 
@@ -37,19 +25,6 @@ export type LoadedViewport = {
 export type ListViewport = {
   top: number;
   bottom: number;
-};
-
-export type FileListRenderItemProps = RenderItemProps<FileStorageItem> & {
-  children?: JSX.Element;
-  dropTargetItem?: FileStorageItem;
-  draggedItems?: FileStorageItem[];
-  isDragInProgress: boolean;
-  isDropTargetValid: boolean;
-
-  onDragStart(index: number, e: React.DragEvent<HTMLDivElement>): void;
-  onDragOver(index: number, e: React.DragEvent<HTMLDivElement>): void;
-  onDragEnd(index: number, e: React.DragEvent<HTMLDivElement>): void;
-  onDrop(index: number, e: React.DragEvent<HTMLDivElement>): void;
 };
 
 export interface FileListProps {
@@ -76,104 +51,6 @@ const DRAG_HOVER_TIMEOUT = 500;
 
 const ITEM_LIST_CLASS_NAME = 'item-list-scroll-pane';
 
-export function RenderFileListItem(
-  props: FileListRenderItemProps
-): JSX.Element {
-  const {
-    children,
-    draggedItems,
-    isDragInProgress,
-    isDropTargetValid,
-    isSelected,
-    item,
-    itemIndex,
-    dropTargetItem,
-    onDragStart,
-    onDragOver,
-    onDragEnd,
-    onDrop,
-  } = props;
-
-  const isDragged =
-    draggedItems?.some(draggedItem => draggedItem.id === item.id) ?? false;
-  const itemPath = getPathFromItem(item);
-  const dropTargetPath =
-    isDragInProgress && dropTargetItem ? getPathFromItem(dropTargetItem) : null;
-
-  const isExactDropTarget =
-    isDragInProgress &&
-    isDropTargetValid &&
-    isDirectory(item) &&
-    dropTargetPath === itemPath;
-  const isInDropTarget =
-    isDragInProgress && isDropTargetValid && dropTargetPath === itemPath;
-  const isInvalidDropTarget =
-    isDragInProgress && !isDropTargetValid && dropTargetPath === itemPath;
-
-  const icon = getItemIcon(item);
-  const depth = FileUtils.getDepth(item.filename);
-  const depthLines = Array(depth)
-    .fill(null)
-    .map((value, index) => (
-      // eslint-disable-next-line react/no-array-index-key
-      <span className="file-list-depth-line" key={index} />
-    ));
-
-  return (
-    <div
-      className={classNames(
-        'd-flex w-100 align-items-center',
-        'file-list-item',
-        {
-          'is-dragged': isDragged,
-          'is-exact-drop-target': isExactDropTarget,
-          'is-in-drop-target': isInDropTarget,
-          'is-invalid-drop-target': isInvalidDropTarget,
-          'is-selected': isSelected,
-        }
-      )}
-      onDragStart={e => onDragStart(itemIndex, e)}
-      onDragOver={e => onDragOver(itemIndex, e)}
-      onDragEnd={e => onDragEnd(itemIndex, e)}
-      onDrop={e => onDrop(itemIndex, e)}
-      draggable
-      role="presentation"
-      aria-label={item.basename}
-    >
-      {depthLines}{' '}
-      <FontAwesomeIcon icon={icon} className="item-icon" fixedWidth />{' '}
-      <span className="truncation-wrapper">
-        {children ?? item.basename}
-        <Tooltip
-          options={{
-            placement: 'left',
-          }}
-        >
-          {children ?? item.basename}
-        </Tooltip>
-      </span>
-    </div>
-  );
-}
-
-/**
- * Get the icon definition for a file or folder item
- * @param item Item to get the icon for
- * @returns Icon definition to pass in the FontAwesomeIcon icon prop
- */
-function getItemIcon(item: FileStorageItem): IconDefinition {
-  if (isDirectory(item)) {
-    return item.isExpanded ? vsFolderOpened : vsFolder;
-  }
-  const mimeType = FileUtils.getMimeType(item.basename);
-  switch (mimeType) {
-    case MIME_TYPE.PYTHON:
-      return dhPython;
-    default:
-      return vsCode;
-  }
-}
-
 /**
  * Component that displays and allows interaction with the file system in the provided FileStorageTable.
  */
@@ -185,7 +62,7 @@ export function FileList(props: FileListProps): JSX.Element {
     onMove,
     onSelect,
     onSelectionChange = () => undefined,
-    renderItem = RenderFileListItem,
+    renderItem = FileListItem,
     rowHeight = DEFAULT_ROW_HEIGHT,
     overscanCount = ItemList.DEFAULT_OVERSCAN,
   } = props;
